mainとappを修正（再度）
