+++ app.py
@@ -1,119 +1,119 @@
-from fastapi import FastAPI
-from fastapi.middleware.cors import CORSMiddleware
-from app.quickdiagnose import router as quickdiagnose_router
-
-app = FastAPI()
-
-# CORS設定（Next.jsなどからの呼び出し対応）
-origins = [
-    "http://localhost:3000",
-    "http://127.0.0.1:3000"
-]
-
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=origins,
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
-
-# クイック診断機能ルーターの登録
-app.include_router(quickdiagnose_router)
-
-
-
-# #以下Prcticalのオリジナル
-from fastapi import FastAPI, HTTPException, Query
-from fastapi.middleware.cors import CORSMiddleware
-from pydantic import BaseModel
-import requests
-import json
-from db_control import crud, mymodels
-
-# MySQLのテーブル作成
-from db_control.create_tables import init_db
-
-# # アプリケーション初期化時にテーブルを作成
-init_db()
-
-
-class Customer(BaseModel):
-    customer_id: str
-    customer_name: str
-    age: int
-    gender: str
-
+# from fastapi import FastAPI
+# from fastapi.middleware.cors import CORSMiddleware
+# from app.quickdiagnose import router as quickdiagnose_router
 
 # app = FastAPI()
 
-# CORSミドルウェアの設定
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["*"],
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
+# # CORS設定（Next.jsなどからの呼び出し対応）
+# origins = [
+#     "http://localhost:3000",
+#     "http://127.0.0.1:3000"
+# ]
+
+# app.add_middleware(
+#     CORSMiddleware,
+#     allow_origins=origins,
+#     allow_credentials=True,
+#     allow_methods=["*"],
+#     allow_headers=["*"],
+# )
+
+# # クイック診断機能ルーターの登録
+# app.include_router(quickdiagnose_router)
 
 
-@app.get("/")
-def index():
-    return {"message": "FastAPI top page!"}
+
+# # #以下Prcticalのオリジナル
+# from fastapi import FastAPI, HTTPException, Query
+# from fastapi.middleware.cors import CORSMiddleware
+# from pydantic import BaseModel
+# import requests
+# import json
+# from db_control import crud, mymodels
+
+# # MySQLのテーブル作成
+# from db_control.create_tables import init_db
+
+# # # アプリケーション初期化時にテーブルを作成
+# init_db()
 
 
-@app.post("/customers")
-def create_customer(customer: Customer):
-    values = customer.dict()
-    tmp = crud.myinsert(mymodels.Customers, values)
-    result = crud.myselect(mymodels.Customers, values.get("customer_id"))
-
-    if result:
-        result_obj = json.loads(result)
-        return result_obj if result_obj else None
-    return None
+# class Customer(BaseModel):
+#     customer_id: str
+#     customer_name: str
+#     age: int
+#     gender: str
 
 
-@app.get("/customers")
-def read_one_customer(customer_id: str = Query(...)):
-    result = crud.myselect(mymodels.Customers, customer_id)
-    if not result:
-        raise HTTPException(status_code=404, detail="Customer not found")
-    result_obj = json.loads(result)
-    return result_obj[0] if result_obj else None
+# # app = FastAPI()
+
+# # CORSミドルウェアの設定
+# app.add_middleware(
+#     CORSMiddleware,
+#     allow_origins=["*"],
+#     allow_credentials=True,
+#     allow_methods=["*"],
+#     allow_headers=["*"],
+# )
 
 
-@app.get("/allcustomers")
-def read_all_customer():
-    result = crud.myselectAll(mymodels.Customers)
-    # 結果がNoneの場合は空配列を返す
-    if not result:
-        return []
-    # JSON文字列をPythonオブジェクトに変換
-    return json.loads(result)
+# @app.get("/")
+# def index():
+#     return {"message": "FastAPI top page!"}
 
 
-@app.put("/customers")
-def update_customer(customer: Customer):
-    values = customer.dict()
-    values_original = values.copy()
-    tmp = crud.myupdate(mymodels.Customers, values)
-    result = crud.myselect(mymodels.Customers, values_original.get("customer_id"))
-    if not result:
-        raise HTTPException(status_code=404, detail="Customer not found")
-    result_obj = json.loads(result)
-    return result_obj[0] if result_obj else None
+# @app.post("/customers")
+# def create_customer(customer: Customer):
+#     values = customer.dict()
+#     tmp = crud.myinsert(mymodels.Customers, values)
+#     result = crud.myselect(mymodels.Customers, values.get("customer_id"))
+
+#     if result:
+#         result_obj = json.loads(result)
+#         return result_obj if result_obj else None
+#     return None
 
 
-@app.delete("/customers")
-def delete_customer(customer_id: str = Query(...)):
-    result = crud.mydelete(mymodels.Customers, customer_id)
-    if not result:
-        raise HTTPException(status_code=404, detail="Customer not found")
-    return {"customer_id": customer_id, "status": "deleted"}
+# @app.get("/customers")
+# def read_one_customer(customer_id: str = Query(...)):
+#     result = crud.myselect(mymodels.Customers, customer_id)
+#     if not result:
+#         raise HTTPException(status_code=404, detail="Customer not found")
+#     result_obj = json.loads(result)
+#     return result_obj[0] if result_obj else None
 
 
-@app.get("/fetchtest")
-def fetchtest():
-    response = requests.get('https://jsonplaceholder.typicode.com/users')
-    return response.json()
+# @app.get("/allcustomers")
+# def read_all_customer():
+#     result = crud.myselectAll(mymodels.Customers)
+#     # 結果がNoneの場合は空配列を返す
+#     if not result:
+#         return []
+#     # JSON文字列をPythonオブジェクトに変換
+#     return json.loads(result)
+
+
+# @app.put("/customers")
+# def update_customer(customer: Customer):
+#     values = customer.dict()
+#     values_original = values.copy()
+#     tmp = crud.myupdate(mymodels.Customers, values)
+#     result = crud.myselect(mymodels.Customers, values_original.get("customer_id"))
+#     if not result:
+#         raise HTTPException(status_code=404, detail="Customer not found")
+#     result_obj = json.loads(result)
+#     return result_obj[0] if result_obj else None
+
+
+# @app.delete("/customers")
+# def delete_customer(customer_id: str = Query(...)):
+#     result = crud.mydelete(mymodels.Customers, customer_id)
+#     if not result:
+#         raise HTTPException(status_code=404, detail="Customer not found")
+#     return {"customer_id": customer_id, "status": "deleted"}
+
+
+# @app.get("/fetchtest")
+# def fetchtest():
+#     response = requests.get('https://jsonplaceholder.typicode.com/users')
+#     return response.json()
